Update ruff to v0.15.12 (#7439)
All changes here were done with ruff autofix feature, with --unsafe-fixes enabled. I verified that the 'safety' prerequisites were met in each individual case.

+++ src/aiida/restapi/common/utils.py
@@ -277,7 +277,7 @@ class Utils:
         if total_count == 0:
             last_page = 1
         else:
-            last_page = int(ceil(total_count / perpage))
+            last_page = ceil(total_count / perpage)
 
         ## Check validity of required page and calculate limit, offset,
         # previous,
